Adds interface and refactors package
