+++ src/main/groovy/life/qbic/portal/sampletracking/app/samples/ChangeLocation.groovy
@@ -1,7 +1,9 @@
-package life.qbic.portal.sampletracking.app.applychange
+package life.qbic.portal.sampletracking.app.samples
 
 import groovy.util.logging.Log4j2
 
 @Log4j2
 class ChangeLocation {
+
+
 }
