+++ src/main/resources/life/qbic/portal/sampletracking/AppWidgetSet.gwt.xml
@@ -9,4 +9,6 @@
     <!-- Inherit DefaultWidgetSet -->
     <inherits name="com.vaadin.DefaultWidgetSet" /> 
 
+
+    <inherits name="life.qbic.portal.portlet.AppWidgetSet" />
 </module>
